+++ internal/provider/snapmirror_resource_test.go
@@ -34,6 +34,15 @@ func TestAccSnapmirrorResource(t *testing.T) {
 					resource.TestCheckResourceAttr("netapp-ontap_snapmirror_resource.example", "destination_endpoint.path", "snapmirror_source_svm:snap"),
 				),
 			},
+			// Import and read
+			{
+				ResourceName:  "netapp-ontap_snapmirror_resource.example",
+				ImportState:   true,
+				ImportStateId: fmt.Sprintf("%s,%s", "snapmirror_dest_svm:snap_dest", "cluster4"),
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr("netapp-ontap_snapmirror_resource.example", "destination_endpoint.path", "snapmirror_dest_svm:snap_dest"),
+				),
+			},
 		},
 	})
 }